tfmodel/requirements.txt
numpy
pandas
tensorflow
pymongo
kafka-python